# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CampusSlot v1.0 - Priority-Based Lab Management System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export interface SuccessResponse {
  success: boolean;
  message: string;
}

export type Priority = (typeof Priority)[keyof typeof Priority];

export const Priority = {
  P1: "P1",
  P2: "P2",
  P3: "P3",
} as const;

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  PlacementCoordinator: "PlacementCoordinator",
  Faculty: "Faculty",
  LabAssistant: "LabAssistant",
  Admin: "Admin",
  Student: "Student",
} as const;

export interface TimeSlot {
  day: string;
  startTime: string;
  endTime: string;
}

export interface Lab {
  id: string;
  name: string;
  building: string;
  floor: number;
  capacity: number;
  hasI7Processors: boolean;
  hasGraphicsCards: boolean;
  isUnderMaintenance: boolean;
}

export interface Booking {
  id: string;
  subject: string;
  faculty: string;
  labName: string;
  labId: string;
  timeSlot: TimeSlot;
  capacity: number;
  priority: Priority;
  isDisplaced: boolean;
  createdAt: string;
}

export interface BookingResult {
  booking: Booking;
  displaced?: Booking;
  message: string;
}

export interface CreateBookingRequest {
  subject: string;
  faculty: string;
  labId: string;
  timeSlot: TimeSlot;
  capacity: number;
  priority: Priority;
  requiresI7?: boolean;
  requiresGraphics?: boolean;
}

export interface BulkBookingRequest {
  bookings: CreateBookingRequest[];
}

export interface BulkBookingResult {
  created: Booking[];
  displaced: Booking[];
  errors: string[];
}

export interface UpdateLabRequest {
  isUnderMaintenance?: boolean;
  capacity?: number;
  hasI7Processors?: boolean;
  hasGraphicsCards?: boolean;
}

export interface CreateLabRequest {
  name: string;
  building: string;
  floor: number;
  capacity: number;
  hasI7Processors: boolean;
  hasGraphicsCards: boolean;
}

export interface ConflictResolution {
  id: string;
  resolvedAt: string;
  winningBooking: Booking;
  displacedBooking: Booking;
  resolution: string;
}

export interface SuggestionRequest {
  bookingId?: string;
  capacity: number;
  requiresI7?: boolean;
  requiresGraphics?: boolean;
  preferredWeek?: string;
}

export interface SlotSuggestion {
  lab: Lab;
  timeSlot: TimeSlot;
  score: number;
  reason: string;
}

export interface PeakHour {
  hour: string;
  count: number;
}

export interface AnalyticsOverview {
  totalBookings: number;
  p1Count: number;
  p2Count: number;
  p3Count: number;
  totalDisplacements: number;
  peakHours: PeakHour[];
}

export interface LabUtilization {
  labName: string;
  utilizationPercent: number;
  bookingCount: number;
}

export interface AuditLogEntry {
  id: string;
  action: string;
  description: string;
  timestamp: string;
  actor: string;
  subject: string;
  priority: Priority;
  labName?: string;
}

export interface AuditLogResponse {
  entries: AuditLogEntry[];
  totalEntries: number;
  p1Overrides: number;
  systemActions: number;
  manualActions: number;
}

export interface DashboardStats {
  totalLabs: number;
  totalCapacity: number;
  occupancyPercent: number;
  labsInUse: number;
  pendingConflicts: number;
  autoResolved: number;
  availableNow: number;
}

export interface DisplacementAlert {
  id: string;
  bookingId: string;
  subject: string;
  originalLab: string;
  originalSlot: TimeSlot;
  displacedAt: string;
  reason: string;
  isRead: boolean;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export interface RegisterRequest {
  name: string;
  email: string;
  password: string;
  role: UserRole;
  department: string;
  registrationNumber?: string;
}

export interface User {
  id: string;
  name: string;
  email: string;
  role: UserRole;
  department: string;
}

export interface AuthResponse {
  user: User;
  token: string;
  message: string;
}

export type GetBookingsParams = {
  labId?: string;
  faculty?: string;
  week?: string;
};

export type GetAuditLogsParams = {
  search?: string;
  priority?: string;
};