# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: CampusSlot v1.0 - Priority-Based Lab Management System API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication operations
  - name: labs
    description: Lab management
  - name: bookings
    description: Booking management
  - name: conflicts
    description: Conflict resolution
  - name: analytics
    description: Analytics and reporting
  - name: audit
    description: Audit log operations
  - name: suggestions
    description: Smart slot suggestions
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Registration successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  /labs:
    get:
      operationId: getLabs
      tags: [labs]
      summary: Get all labs
      responses:
        "200":
          description: List of labs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lab"
    post:
      operationId: createLab
      tags: [labs]
      summary: Create a lab
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLabRequest"
      responses:
        "201":
          description: Lab created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lab"

  /labs/{labId}:
    patch:
      operationId: updateLab
      tags: [labs]
      summary: Update lab (maintenance toggle etc)
      parameters:
        - name: labId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLabRequest"
      responses:
        "200":
          description: Lab updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lab"

  /bookings:
    get:
      operationId: getBookings
      tags: [bookings]
      summary: Get all bookings
      parameters:
        - name: labId
          in: query
          schema:
            type: string
        - name: faculty
          in: query
          schema:
            type: string
        - name: week
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
    post:
      operationId: createBooking
      tags: [bookings]
      summary: Create a booking (with conflict resolution)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookingRequest"
      responses:
        "201":
          description: Booking created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingResult"
        "409":
          description: Conflict - equal priority booking exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /bookings/bulk:
    post:
      operationId: createBulkBookings
      tags: [bookings]
      summary: Create multiple bookings at once
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkBookingRequest"
      responses:
        "201":
          description: Bookings created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkBookingResult"

  /bookings/{bookingId}:
    delete:
      operationId: deleteBooking
      tags: [bookings]
      summary: Delete a booking
      parameters:
        - name: bookingId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Booking deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /conflicts:
    get:
      operationId: getConflicts
      tags: [conflicts]
      summary: Get all conflict resolutions
      responses:
        "200":
          description: List of conflict resolutions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ConflictResolution"

  /suggestions:
    post:
      operationId: getSuggestions
      tags: [suggestions]
      summary: Get smart slot suggestions for displaced booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SuggestionRequest"
      responses:
        "200":
          description: List of slot suggestions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SlotSuggestion"

  /analytics/overview:
    get:
      operationId: getAnalyticsOverview
      tags: [analytics]
      summary: Get analytics overview
      responses:
        "200":
          description: Analytics overview data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyticsOverview"

  /analytics/utilization:
    get:
      operationId: getUtilizationData
      tags: [analytics]
      summary: Get lab utilization rates
      responses:
        "200":
          description: Utilization data per lab
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LabUtilization"

  /audit:
    get:
      operationId: getAuditLogs
      tags: [audit]
      summary: Get audit logs
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: priority
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Audit logs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditLogResponse"

  /dashboard/stats:
    get:
      operationId: getDashboardStats
      tags: [analytics]
      summary: Get dashboard statistics
      responses:
        "200":
          description: Dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"

  /alerts:
    get:
      operationId: getAlerts
      tags: [bookings]
      summary: Get displacement alerts for current faculty
      responses:
        "200":
          description: Faculty alerts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DisplacementAlert"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message

    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message

    Priority:
      type: string
      enum: [P1, P2, P3]

    UserRole:
      type: string
      enum: [PlacementCoordinator, Faculty, LabAssistant, Admin, Student]

    TimeSlot:
      type: object
      properties:
        day:
          type: string
        startTime:
          type: string
        endTime:
          type: string
      required:
        - day
        - startTime
        - endTime

    Lab:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        building:
          type: string
        floor:
          type: integer
        capacity:
          type: integer
        hasI7Processors:
          type: boolean
        hasGraphicsCards:
          type: boolean
        isUnderMaintenance:
          type: boolean
      required:
        - id
        - name
        - building
        - floor
        - capacity
        - hasI7Processors
        - hasGraphicsCards
        - isUnderMaintenance

    Booking:
      type: object
      properties:
        id:
          type: string
        subject:
          type: string
        faculty:
          type: string
        labName:
          type: string
        labId:
          type: string
        timeSlot:
          $ref: "#/components/schemas/TimeSlot"
        capacity:
          type: integer
        priority:
          $ref: "#/components/schemas/Priority"
        isDisplaced:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - subject
        - faculty
        - labName
        - labId
        - timeSlot
        - capacity
        - priority
        - isDisplaced
        - createdAt

    BookingResult:
      type: object
      properties:
        booking:
          $ref: "#/components/schemas/Booking"
        displaced:
          $ref: "#/components/schemas/Booking"
        message:
          type: string
      required:
        - booking
        - message

    BulkBookingRequest:
      type: object
      properties:
        bookings:
          type: array
          items:
            $ref: "#/components/schemas/CreateBookingRequest"
      required:
        - bookings

    BulkBookingResult:
      type: object
      properties:
        created:
          type: array
          items:
            $ref: "#/components/schemas/Booking"
        displaced:
          type: array
          items:
            $ref: "#/components/schemas/Booking"
        errors:
          type: array
          items:
            type: string
      required:
        - created
        - displaced
        - errors

    CreateBookingRequest:
      type: object
      properties:
        subject:
          type: string
        faculty:
          type: string
        labId:
          type: string
        timeSlot:
          $ref: "#/components/schemas/TimeSlot"
        capacity:
          type: integer
        priority:
          $ref: "#/components/schemas/Priority"
        requiresI7:
          type: boolean
        requiresGraphics:
          type: boolean
      required:
        - subject
        - faculty
        - labId
        - timeSlot
        - capacity
        - priority

    UpdateLabRequest:
      type: object
      properties:
        isUnderMaintenance:
          type: boolean
        capacity:
          type: integer
        hasI7Processors:
          type: boolean
        hasGraphicsCards:
          type: boolean

    CreateLabRequest:
      type: object
      properties:
        name:
          type: string
        building:
          type: string
        floor:
          type: integer
        capacity:
          type: integer
        hasI7Processors:
          type: boolean
        hasGraphicsCards:
          type: boolean
      required:
        - name
        - building
        - floor
        - capacity
        - hasI7Processors
        - hasGraphicsCards

    ConflictResolution:
      type: object
      properties:
        id:
          type: string
        resolvedAt:
          type: string
        winningBooking:
          $ref: "#/components/schemas/Booking"
        displacedBooking:
          $ref: "#/components/schemas/Booking"
        resolution:
          type: string
      required:
        - id
        - resolvedAt
        - winningBooking
        - displacedBooking
        - resolution

    SuggestionRequest:
      type: object
      properties:
        bookingId:
          type: string
        capacity:
          type: integer
        requiresI7:
          type: boolean
        requiresGraphics:
          type: boolean
        preferredWeek:
          type: string
      required:
        - capacity

    SlotSuggestion:
      type: object
      properties:
        lab:
          $ref: "#/components/schemas/Lab"
        timeSlot:
          $ref: "#/components/schemas/TimeSlot"
        score:
          type: integer
        reason:
          type: string
      required:
        - lab
        - timeSlot
        - score
        - reason

    AnalyticsOverview:
      type: object
      properties:
        totalBookings:
          type: integer
        p1Count:
          type: integer
        p2Count:
          type: integer
        p3Count:
          type: integer
        totalDisplacements:
          type: integer
        peakHours:
          type: array
          items:
            $ref: "#/components/schemas/PeakHour"
      required:
        - totalBookings
        - p1Count
        - p2Count
        - p3Count
        - totalDisplacements
        - peakHours

    PeakHour:
      type: object
      properties:
        hour:
          type: string
        count:
          type: integer
      required:
        - hour
        - count

    LabUtilization:
      type: object
      properties:
        labName:
          type: string
        utilizationPercent:
          type: integer
        bookingCount:
          type: integer
      required:
        - labName
        - utilizationPercent
        - bookingCount

    AuditLogEntry:
      type: object
      properties:
        id:
          type: string
        action:
          type: string
        description:
          type: string
        timestamp:
          type: string
        actor:
          type: string
        subject:
          type: string
        priority:
          $ref: "#/components/schemas/Priority"
        labName:
          type: string
      required:
        - id
        - action
        - description
        - timestamp
        - actor
        - subject
        - priority

    AuditLogResponse:
      type: object
      properties:
        entries:
          type: array
          items:
            $ref: "#/components/schemas/AuditLogEntry"
        totalEntries:
          type: integer
        p1Overrides:
          type: integer
        systemActions:
          type: integer
        manualActions:
          type: integer
      required:
        - entries
        - totalEntries
        - p1Overrides
        - systemActions
        - manualActions

    DashboardStats:
      type: object
      properties:
        totalLabs:
          type: integer
        totalCapacity:
          type: integer
        occupancyPercent:
          type: integer
        labsInUse:
          type: integer
        pendingConflicts:
          type: integer
        autoResolved:
          type: integer
        availableNow:
          type: integer
      required:
        - totalLabs
        - totalCapacity
        - occupancyPercent
        - labsInUse
        - pendingConflicts
        - autoResolved
        - availableNow

    DisplacementAlert:
      type: object
      properties:
        id:
          type: string
        bookingId:
          type: string
        subject:
          type: string
        originalLab:
          type: string
        originalSlot:
          $ref: "#/components/schemas/TimeSlot"
        displacedAt:
          type: string
        reason:
          type: string
        isRead:
          type: boolean
      required:
        - id
        - bookingId
        - subject
        - originalLab
        - originalSlot
        - displacedAt
        - reason
        - isRead

    LoginRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password

    RegisterRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          $ref: "#/components/schemas/UserRole"
        department:
          type: string
        registrationNumber:
          type: string
      required:
        - name
        - email
        - password
        - role
        - department

    User:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        role:
          $ref: "#/components/schemas/UserRole"
        department:
          type: string
      required:
        - id
        - name
        - email
        - role
        - department

    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
        message:
          type: string
      required:
        - user
        - token
        - message